code clean

+++ mindspore/core/ops/csr_sparse_matrix_to_dense.cc
@@ -83,7 +83,7 @@ abstract::ShapePtr CSRSparseMatrixToDenseInferShape(const PrimitivePtr &primitiv
   auto d_shape_value_ptr_tensor =
     CheckAndConvertUtils::CheckTensorIntValue("x_dense_shape", d_shape_value_ptr, primitive->name());
   for (int64_t i = kZero; i < rank; i++) {
-    if (static_cast<int64_t>(d_shape_value_ptr_tensor[i]) <= kZero) {
+    if (static_cast<int64_t>(d_shape_value_ptr_tensor[static_cast<size_t>(i)]) <= kZero) {
       MS_EXCEPTION(ValueError) << "For '" << primitive->name()
                                << "', each element of 'x_dense_shape' must be greater than 0.";
     }
