+++ mindspore/core/ops/grad/sparse_segment_mean_grad.cc
@@ -93,7 +93,7 @@ AbstractBasePtr SparseSegmentMeanGradInfer(const abstract::AnalysisEnginePtr &, 
                                            const std::vector<AbstractBasePtr> &input_args) {
   MS_EXCEPTION_IF_NULL(prim);
   auto prim_name = prim->name();
-  const int64_t input_num = kInputIndex4;
+  const int64_t input_num = static_cast<int64_t>(kInputIndex4);
   CheckAndConvertUtils::CheckInputArgs(input_args, kEqual, input_num, prim_name);
   auto types = SparseSegmentMeanGradInferType(prim, input_args);
   auto shapes = SparseSegmentMeanGradInferShape(prim, input_args);
